Moved tests to their own file
